Added the full dashboard with all the styles and the backend code to get the security scores of each security subject and the notifications count.

+++ api/src/main/java/com/cloudmen/cloudguard/dto/users/UserOverviewResponse.java
@@ -1,10 +1,10 @@
 package com.cloudmen.cloudguard.dto.users;
 
 public record UserOverviewResponse(
-            long totalUsers,
-            long withoutTwoFactor,
-            long adminUsers,
-            long securityScore,
-            long activeLongNoLoginCount,
-            long inactiveRecentLoginCount
+            int totalUsers,
+            int withoutTwoFactor,
+            int adminUsers,
+            int securityScore,
+            int activeLongNoLoginCount,
+            int inactiveRecentLoginCount
 ){}
